compiler warning fixed
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@3639 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ RenderDevice.cpp
@@ -257,7 +257,7 @@ void EnumerateDisplayModes(const int display, std::vector<VideoMode>& modes)
 
    std::vector<DisplayConfig> displays;
    getDisplayList(displays);
-   if (display >= displays.size())
+   if (display >= (int)displays.size())
       return;
    const int adapter = displays[display].adapter;
